refactor(tools): remove dedicated search tools
