+++ src/tool-surfaces/claude.ts
@@ -1,0 +1,317 @@
+import { registerAppTool } from "@modelcontextprotocol/ext-apps/server";
+import { existsSync } from "node:fs";
+import * as z from "zod/v4";
+import {
+  editFileTool,
+  runShellTool,
+  writeFileTool,
+} from "../pi-tools.js";
+import {
+  EDIT_TOOL_ANNOTATIONS,
+  SHELL_TOOL_ANNOTATIONS,
+  WRITE_TOOL_ANNOTATIONS,
+  toolNames,
+  workspaceIdDescription,
+  type ToolInstructionContext,
+  type ToolRegistrationContext,
+} from "./types.js";
+import {
+  contentLineCount,
+  contentText,
+  countDiffStats,
+  logFailedToolResponse,
+  logToolCall,
+  newFilePatch,
+  resultOutputSchema,
+  textBlock,
+  textSummary,
+  toolWidgetDescriptorMeta,
+} from "./shared.js";
+
+const CLAUDE_INSTRUCTIONS = `Use ${toolNames.read} for direct file reads, ${toolNames.shell} with command-line tools such as rg, find, ls, and tree for search and directory inspection, ${toolNames.edit} for targeted modifications, and ${toolNames.write} only for new files or complete rewrites. Use ${toolNames.shell} for tests, builds, git inspection, package scripts, and other commands, but do not create or modify files through shell commands. Follow instructions returned by ${toolNames.openWorkspace}; read applicable instruction and skill files before working in their scope.`;
+
+export function claudeInstructions({
+  agents,
+  skills,
+}: ToolInstructionContext): string {
+  return `${agents}${skills}${CLAUDE_INSTRUCTIONS}`;
+}
+
+export function registerClaudeTools(context: ToolRegistrationContext): void {
+  registerClaudeMutationTools(context);
+  registerShellTool(context);
+}
+
+const CLAUDE_SHELL_DESCRIPTION = `Run a shell command in a workspace. Use it for tests, builds, git inspection, package scripts, search, file discovery, and directory inspection. Do not use ${toolNames.shell} to create or modify files. Do not use shell redirection, heredocs, tee, sed -i, perl -i, node/python/ruby scripts, or generated scripts to write project files; use ${toolNames.edit} for targeted changes and ${toolNames.write} for new files or full rewrites. Prefer ${toolNames.read} for direct file reads. This is powerful execution and should only be exposed behind strong authentication.`;
+
+function registerClaudeMutationTools(context: ToolRegistrationContext): void {
+  const { server, config, workspaces } = context;
+
+  registerAppTool(
+    server,
+    toolNames.write,
+    {
+      title: "Write file",
+      description: `Create or completely overwrite a file in a workspace. Prefer ${toolNames.edit} for targeted changes to existing files.`,
+      inputSchema: {
+        workspaceId: z.string().describe(workspaceIdDescription),
+        path: z
+          .string()
+          .describe("File path to write, relative to the workspace root."),
+        content: z.string().describe("Complete new file content."),
+      },
+      outputSchema: resultOutputSchema(),
+      ...toolWidgetDescriptorMeta(config, "write"),
+      annotations: WRITE_TOOL_ANNOTATIONS,
+    },
+    async ({ workspaceId, ...input }) => {
+      const startedAt = performance.now();
+      const workspace = workspaces.getWorkspace(workspaceId);
+      const absolutePath = workspaces.resolvePath(workspace, input.path);
+      const overwritesExistingFile = existsSync(absolutePath);
+      const response = await writeFileTool(input, {
+        cwd: workspace.root,
+        root: workspace.root,
+      });
+
+      if (response.isError) {
+        logFailedToolResponse(
+          config,
+          {
+            tool: toolNames.write,
+            workspaceId,
+            path: input.path,
+          },
+          response.content,
+          startedAt,
+        );
+        return response;
+      }
+
+      // An aggregate review can show the real replacement diff. A new-file
+      // patch would misrepresent an overwrite as additions with no removals.
+      const patch = overwritesExistingFile
+        ? undefined
+        : newFilePatch(input.path, input.content);
+      const stats = countDiffStats(patch);
+      const summary = {
+        ...stats,
+        lines: contentLineCount(input.content),
+        characters: input.content.length,
+      };
+      logToolCall(config, {
+        tool: toolNames.write,
+        workspaceId,
+        path: input.path,
+        success: true,
+        durationMs: Math.round(performance.now() - startedAt),
+      });
+
+      return {
+        ...response,
+        _meta: {
+          tool: toolNames.write,
+          card: {
+            workspaceId,
+            path: input.path,
+            summary,
+            payload: {
+              content: response.content,
+              patch,
+            },
+          },
+        },
+        structuredContent: {
+          result: contentText(response.content),
+        },
+      };
+    },
+  );
+
+  registerAppTool(
+    server,
+    toolNames.edit,
+    {
+      title: "Edit file",
+      description: `Edit one file in a workspace by replacing exact text blocks. Prefer this over ${toolNames.write} for targeted changes. Each oldText must match a unique, non-overlapping region of the original file; merge nearby changes into one edit and keep oldText as small as possible while still unique.`,
+      inputSchema: {
+        workspaceId: z.string().describe(workspaceIdDescription),
+        path: z
+          .string()
+          .describe("File path to edit, relative to the workspace root."),
+        edits: z
+          .array(
+            z.object({
+              oldText: z
+                .string()
+                .describe(
+                  "Exact text to replace. Must match uniquely in the original file.",
+                ),
+              newText: z.string().describe("Replacement text."),
+            }),
+          )
+          .min(1),
+      },
+      outputSchema: resultOutputSchema({
+        status: z.literal("applied"),
+      }),
+      ...toolWidgetDescriptorMeta(config, "edit"),
+      annotations: EDIT_TOOL_ANNOTATIONS,
+    },
+    async ({ workspaceId, ...input }) => {
+      const startedAt = performance.now();
+      const workspace = workspaces.getWorkspace(workspaceId);
+      workspaces.resolvePath(workspace, input.path);
+      const response = await editFileTool(input, {
+        cwd: workspace.root,
+        root: workspace.root,
+      });
+
+      if (response.isError) {
+        logFailedToolResponse(
+          config,
+          {
+            tool: toolNames.edit,
+            workspaceId,
+            path: input.path,
+          },
+          response.content,
+          startedAt,
+        );
+        return response;
+      }
+
+      const stats = countDiffStats(
+        response.details?.patch ?? response.details?.diff,
+      );
+      const summary = {
+        ...stats,
+        editCount: input.edits.length,
+      };
+      const editResultText = `Edited ${input.path} (+${stats.additions} -${stats.removals}).`;
+      const editContent = [textBlock(editResultText)];
+      logToolCall(config, {
+        tool: toolNames.edit,
+        workspaceId,
+        path: input.path,
+        success: true,
+        durationMs: Math.round(performance.now() - startedAt),
+      });
+
+      return {
+        content: editContent,
+        _meta: {
+          tool: toolNames.edit,
+          card: {
+            workspaceId,
+            path: input.path,
+            summary,
+            payload: {
+              diff: response.details?.diff,
+              patch: response.details?.patch,
+            },
+          },
+        },
+        structuredContent: {
+          status: "applied",
+          result: contentText(editContent),
+        },
+      };
+    },
+  );
+}
+
+function registerShellTool(context: ToolRegistrationContext): void {
+  const { server, config, workspaces } = context;
+
+  registerAppTool(
+    server,
+    toolNames.shell,
+    {
+      title: "Bash",
+      description: CLAUDE_SHELL_DESCRIPTION,
+      inputSchema: {
+        workspaceId: z.string().describe(workspaceIdDescription),
+        command: z
+          .string()
+          .describe(
+            `Shell command to run. Must not create or modify project files; use ${toolNames.edit} or ${toolNames.write} for file changes.`,
+          ),
+        workingDirectory: z
+          .string()
+          .optional()
+          .describe(
+            "Optional working directory relative to the workspace root. Defaults to the workspace root.",
+          ),
+        timeout: z
+          .number()
+          .positive()
+          .max(300)
+          .optional()
+          .describe("Timeout in seconds. Defaults to 30, max 300."),
+      },
+      outputSchema: resultOutputSchema(),
+      ...toolWidgetDescriptorMeta(config, "shell"),
+      annotations: SHELL_TOOL_ANNOTATIONS,
+    },
+    async ({ workspaceId, workingDirectory, ...input }) => {
+      const startedAt = performance.now();
+      const workspace = workspaces.getWorkspace(workspaceId);
+      const cwd = workspaces.resolveWorkingDirectory(
+        workspace,
+        workingDirectory,
+      );
+      const response = await runShellTool(input, {
+        cwd,
+        root: workspace.root,
+      });
+
+      if (response.isError) {
+        logFailedToolResponse(
+          config,
+          {
+            tool: toolNames.shell,
+            workspaceId,
+            workingDirectory: workingDirectory ?? ".",
+            command: input.command,
+            commandLength: input.command.length,
+          },
+          response.content,
+          startedAt,
+        );
+        return response;
+      }
+
+      const summary = {
+        command: input.command,
+        workingDirectory: workingDirectory ?? ".",
+        ...textSummary(response.content),
+      };
+      logToolCall(config, {
+        tool: toolNames.shell,
+        workspaceId,
+        workingDirectory: workingDirectory ?? ".",
+        command: input.command,
+        commandLength: input.command.length,
+        success: true,
+        durationMs: Math.round(performance.now() - startedAt),
+      });
+
+      return {
+        ...response,
+        _meta: {
+          tool: toolNames.shell,
+          card: {
+            workspaceId,
+            path: workingDirectory,
+            summary,
+            payload: { content: response.content },
+          },
+        },
+        structuredContent: {
+          result: contentText(response.content),
+        },
+      };
+    },
+  );
+}
