--- src/tool-surfaces/standard.ts
@@ -1,573 +1,0 @@
-import { registerAppTool } from "@modelcontextprotocol/ext-apps/server";
-import { existsSync } from "node:fs";
-import * as z from "zod/v4";
-import {
-  editFileTool,
-  findFilesTool,
-  grepFilesTool,
-  listDirectoryTool,
-  runShellTool,
-  writeFileTool,
-} from "../pi-tools.js";
-import {
-  EDIT_TOOL_ANNOTATIONS,
-  SHELL_TOOL_ANNOTATIONS,
-  WRITE_TOOL_ANNOTATIONS,
-  toolNames,
-  workspaceIdDescription,
-  type ToolInstructionContext,
-  type ToolRegistrationContext,
-} from "./types.js";
-import {
-  contentLineCount,
-  contentText,
-  countDiffStats,
-  logFailedToolResponse,
-  logToolCall,
-  newFilePatch,
-  resultOutputSchema,
-  textBlock,
-  textSummary,
-  toolWidgetDescriptorMeta,
-} from "./shared.js";
-
-type StandardRegistration = (context: ToolRegistrationContext) => void;
-
-const MINIMAL_INSPECTION = `In minimal tool mode, ${toolNames.grep}, ${toolNames.glob}, and ${toolNames.ls} are disabled; use ${toolNames.shell} with command-line tools such as grep, rg, find, ls, and tree for search and directory inspection. `;
-
-const FULL_INSPECTION = `Prefer ${toolNames.read}, ${toolNames.grep}, ${toolNames.glob}, and ${toolNames.ls} for file inspection. `;
-
-const STANDARD_EDITING = `Prefer ${toolNames.edit} for targeted modifications, ${toolNames.write} only for new files or complete rewrites, and ${toolNames.shell} for tests, builds, git inspection, package scripts, and commands that are better executed by the shell. Do not create or modify files with ${toolNames.shell}; avoid shell redirection, heredocs, tee, sed -i, perl -i, node/python/ruby scripts, or any command whose purpose is to write project files.`;
-
-export function standardInstructions(mode: "minimal" | "full") {
-  const inspection = mode === "minimal" ? MINIMAL_INSPECTION : FULL_INSPECTION;
-  return ({ agents, skills }: ToolInstructionContext): string =>
-    `${agents}${skills}${inspection}${STANDARD_EDITING}`;
-}
-
-export function registerStandardTools(
-  context: ToolRegistrationContext,
-  mode: "minimal" | "full",
-): void {
-  for (const register of STANDARD_REGISTRATIONS[mode]) {
-    register(context);
-  }
-}
-
-const STANDARD_REGISTRATIONS: Record<
-  "minimal" | "full",
-  readonly StandardRegistration[]
-> = {
-  minimal: [registerStandardMutationTools, registerMinimalShellTool],
-  full: [
-    registerStandardMutationTools,
-    registerSearchTools,
-    registerFullShellTool,
-  ],
-};
-
-const MINIMAL_SHELL_DESCRIPTION = `Run a shell command in a workspace. Use only for tests, builds, git inspection, package scripts, search, file discovery, and directory inspection. In minimal tool mode, ${toolNames.grep}, ${toolNames.glob}, and ${toolNames.ls} are disabled; use command-line tools such as grep, rg, find, ls, and tree for those read-only inspection actions. Do not use ${toolNames.shell} to create or modify files. Do not use shell redirection, heredocs, tee, sed -i, perl -i, node/python/ruby scripts, or generated scripts to write project files; use ${toolNames.edit} for targeted changes and ${toolNames.write} for new files or full rewrites. Prefer ${toolNames.read} for direct file reads. This is powerful execution and should only be exposed behind strong authentication.`;
-const FULL_SHELL_DESCRIPTION = `Run a shell command in a workspace. Use only for tests, builds, git inspection, package scripts, and commands that are better executed by the shell. Do not use ${toolNames.shell} to create or modify files. Do not use shell redirection, heredocs, tee, sed -i, perl -i, node/python/ruby scripts, or generated scripts to write project files; use ${toolNames.edit} for targeted changes and ${toolNames.write} for new files or full rewrites. Prefer ${toolNames.read}, ${toolNames.grep}, ${toolNames.glob}, and ${toolNames.ls} for file inspection. This is powerful execution and should only be exposed behind strong authentication.`;
-
-function registerMinimalShellTool(context: ToolRegistrationContext): void {
-  registerShellTool(context, MINIMAL_SHELL_DESCRIPTION);
-}
-
-function registerFullShellTool(context: ToolRegistrationContext): void {
-  registerShellTool(context, FULL_SHELL_DESCRIPTION);
-}
-
-function registerStandardMutationTools(context: ToolRegistrationContext): void {
-  const { server, config, workspaces } = context;
-
-  registerAppTool(
-    server,
-    toolNames.write,
-    {
-      title: "Write file",
-      description: `Create or completely overwrite a file in a workspace. Prefer ${toolNames.edit} for targeted changes to existing files.`,
-      inputSchema: {
-        workspaceId: z.string().describe(workspaceIdDescription),
-        path: z
-          .string()
-          .describe("File path to write, relative to the workspace root."),
-        content: z.string().describe("Complete new file content."),
-      },
-      outputSchema: resultOutputSchema(),
-      ...toolWidgetDescriptorMeta(config, "write"),
-      annotations: WRITE_TOOL_ANNOTATIONS,
-    },
-    async ({ workspaceId, ...input }) => {
-      const startedAt = performance.now();
-      const workspace = workspaces.getWorkspace(workspaceId);
-      const absolutePath = workspaces.resolvePath(workspace, input.path);
-      const overwritesExistingFile = existsSync(absolutePath);
-      const response = await writeFileTool(input, {
-        cwd: workspace.root,
-        root: workspace.root,
-      });
-
-      if (response.isError) {
-        logFailedToolResponse(
-          config,
-          {
-            tool: toolNames.write,
-            workspaceId,
-            path: input.path,
-          },
-          response.content,
-          startedAt,
-        );
-        return response;
-      }
-
-      // An aggregate review can show the real replacement diff. A new-file
-      // patch would misrepresent an overwrite as additions with no removals.
-      const patch = overwritesExistingFile
-        ? undefined
-        : newFilePatch(input.path, input.content);
-      const stats = countDiffStats(patch);
-      const summary = {
-        ...stats,
-        lines: contentLineCount(input.content),
-        characters: input.content.length,
-      };
-      logToolCall(config, {
-        tool: toolNames.write,
-        workspaceId,
-        path: input.path,
-        success: true,
-        durationMs: Math.round(performance.now() - startedAt),
-      });
-
-      return {
-        ...response,
-        _meta: {
-          tool: toolNames.write,
-          card: {
-            workspaceId,
-            path: input.path,
-            summary,
-            payload: {
-              content: response.content,
-              patch,
-            },
-          },
-        },
-        structuredContent: {
-          result: contentText(response.content),
-        },
-      };
-    },
-  );
-
-  registerAppTool(
-    server,
-    toolNames.edit,
-    {
-      title: "Edit file",
-      description: `Edit one file in a workspace by replacing exact text blocks. Prefer this over ${toolNames.write} for targeted changes. Each oldText must match a unique, non-overlapping region of the original file; merge nearby changes into one edit and keep oldText as small as possible while still unique.`,
-      inputSchema: {
-        workspaceId: z.string().describe(workspaceIdDescription),
-        path: z
-          .string()
-          .describe("File path to edit, relative to the workspace root."),
-        edits: z
-          .array(
-            z.object({
-              oldText: z
-                .string()
-                .describe(
-                  "Exact text to replace. Must match uniquely in the original file.",
-                ),
-              newText: z.string().describe("Replacement text."),
-            }),
-          )
-          .min(1),
-      },
-      outputSchema: resultOutputSchema({
-        status: z.literal("applied"),
-      }),
-      ...toolWidgetDescriptorMeta(config, "edit"),
-      annotations: EDIT_TOOL_ANNOTATIONS,
-    },
-    async ({ workspaceId, ...input }) => {
-      const startedAt = performance.now();
-      const workspace = workspaces.getWorkspace(workspaceId);
-      workspaces.resolvePath(workspace, input.path);
-      const response = await editFileTool(input, {
-        cwd: workspace.root,
-        root: workspace.root,
-      });
-
-      if (response.isError) {
-        logFailedToolResponse(
-          config,
-          {
-            tool: toolNames.edit,
-            workspaceId,
-            path: input.path,
-          },
-          response.content,
-          startedAt,
-        );
-        return response;
-      }
-
-      const stats = countDiffStats(
-        response.details?.patch ?? response.details?.diff,
-      );
-      const summary = {
-        ...stats,
-        editCount: input.edits.length,
-      };
-      const editResultText = `Edited ${input.path} (+${stats.additions} -${stats.removals}).`;
-      const editContent = [textBlock(editResultText)];
-      logToolCall(config, {
-        tool: toolNames.edit,
-        workspaceId,
-        path: input.path,
-        success: true,
-        durationMs: Math.round(performance.now() - startedAt),
-      });
-
-      return {
-        content: editContent,
-        _meta: {
-          tool: toolNames.edit,
-          card: {
-            workspaceId,
-            path: input.path,
-            summary,
-            payload: {
-              diff: response.details?.diff,
-              patch: response.details?.patch,
-            },
-          },
-        },
-        structuredContent: {
-          status: "applied",
-          result: contentText(editContent),
-        },
-      };
-    },
-  );
-}
-
-function registerSearchTools(context: ToolRegistrationContext): void {
-  const { server, config, workspaces } = context;
-
-  registerAppTool(
-    server,
-    toolNames.grep,
-    {
-      title: "Grep",
-      description:
-        "Search file contents in a workspace. Use this before broad reads when looking for symbols, text, or usage sites. Respects project ignore rules.",
-      inputSchema: {
-        workspaceId: z.string().describe(workspaceIdDescription),
-        pattern: z.string().describe("Search pattern."),
-        path: z
-          .string()
-          .optional()
-          .describe(
-            "Optional path or glob scope relative to the workspace root.",
-          ),
-        include: z.string().optional().describe("Optional include glob."),
-      },
-      outputSchema: resultOutputSchema(),
-      ...toolWidgetDescriptorMeta(config, "search"),
-      annotations: { readOnlyHint: true },
-    },
-    async ({ workspaceId, ...input }) => {
-      const startedAt = performance.now();
-      const workspace = workspaces.getWorkspace(workspaceId);
-      if (input.path) workspaces.resolvePath(workspace, input.path);
-      const response = await grepFilesTool(input, {
-        cwd: workspace.root,
-        root: workspace.root,
-      });
-
-      if (response.isError) {
-        logFailedToolResponse(
-          config,
-          {
-            tool: toolNames.grep,
-            workspaceId,
-            path: input.path,
-          },
-          response.content,
-          startedAt,
-        );
-        return response;
-      }
-
-      const summary = {
-        pattern: input.pattern,
-        scope: input.path ?? ".",
-        ...textSummary(response.content),
-      };
-      logToolCall(config, {
-        tool: toolNames.grep,
-        workspaceId,
-        path: input.path,
-        success: true,
-        durationMs: Math.round(performance.now() - startedAt),
-      });
-
-      return {
-        ...response,
-        _meta: {
-          tool: toolNames.grep,
-          card: {
-            workspaceId,
-            path: input.path,
-            summary,
-            payload: { content: response.content },
-          },
-        },
-        structuredContent: {
-          result: contentText(response.content),
-        },
-      };
-    },
-  );
-
-  registerAppTool(
-    server,
-    toolNames.glob,
-    {
-      title: "Glob",
-      description:
-        "Find files by glob pattern in a workspace. Use this to discover filenames or narrow file sets before reading. Respects project ignore rules.",
-      inputSchema: {
-        workspaceId: z.string().describe(workspaceIdDescription),
-        pattern: z.string().describe("File glob pattern."),
-        path: z
-          .string()
-          .optional()
-          .describe("Optional path scope relative to the workspace root."),
-      },
-      outputSchema: resultOutputSchema(),
-      ...toolWidgetDescriptorMeta(config, "search"),
-      annotations: { readOnlyHint: true },
-    },
-    async ({ workspaceId, ...input }) => {
-      const startedAt = performance.now();
-      const workspace = workspaces.getWorkspace(workspaceId);
-      if (input.path) workspaces.resolvePath(workspace, input.path);
-      const response = await findFilesTool(input, {
-        cwd: workspace.root,
-        root: workspace.root,
-      });
-
-      if (response.isError) {
-        logFailedToolResponse(
-          config,
-          {
-            tool: toolNames.glob,
-            workspaceId,
-            path: input.path,
-          },
-          response.content,
-          startedAt,
-        );
-        return response;
-      }
-
-      const summary = {
-        pattern: input.pattern,
-        scope: input.path ?? ".",
-        ...textSummary(response.content),
-      };
-      logToolCall(config, {
-        tool: toolNames.glob,
-        workspaceId,
-        path: input.path,
-        success: true,
-        durationMs: Math.round(performance.now() - startedAt),
-      });
-
-      return {
-        ...response,
-        _meta: {
-          tool: toolNames.glob,
-          card: {
-            workspaceId,
-            path: input.path,
-            summary,
-            payload: { content: response.content },
-          },
-        },
-        structuredContent: {
-          result: contentText(response.content),
-        },
-      };
-    },
-  );
-
-  registerAppTool(
-    server,
-    toolNames.ls,
-    {
-      title: "Ls",
-      description:
-        "List a directory in a workspace. Use this for directory inspection before reading files.",
-      inputSchema: {
-        workspaceId: z.string().describe(workspaceIdDescription),
-        path: z
-          .string()
-          .describe("Directory path to list, relative to the workspace root."),
-      },
-      outputSchema: resultOutputSchema(),
-      ...toolWidgetDescriptorMeta(config, "directory"),
-      annotations: { readOnlyHint: true },
-    },
-    async ({ workspaceId, ...input }) => {
-      const startedAt = performance.now();
-      const workspace = workspaces.getWorkspace(workspaceId);
-      workspaces.resolvePath(workspace, input.path);
-      const response = await listDirectoryTool(input, {
-        cwd: workspace.root,
-        root: workspace.root,
-      });
-
-      if (response.isError) {
-        logFailedToolResponse(
-          config,
-          {
-            tool: toolNames.ls,
-            workspaceId,
-            path: input.path,
-          },
-          response.content,
-          startedAt,
-        );
-        return response;
-      }
-
-      const summary = textSummary(response.content);
-      logToolCall(config, {
-        tool: toolNames.ls,
-        workspaceId,
-        path: input.path,
-        success: true,
-        durationMs: Math.round(performance.now() - startedAt),
-      });
-
-      return {
-        ...response,
-        _meta: {
-          tool: toolNames.ls,
-          card: {
-            workspaceId,
-            path: input.path,
-            summary,
-            payload: { content: response.content },
-          },
-        },
-        structuredContent: {
-          result: contentText(response.content),
-        },
-      };
-    },
-  );
-}
-
-function registerShellTool(
-  context: ToolRegistrationContext,
-  shellDescription: string,
-): void {
-  const { server, config, workspaces } = context;
-
-  registerAppTool(
-    server,
-    toolNames.shell,
-    {
-      title: "Bash",
-      description: shellDescription,
-      inputSchema: {
-        workspaceId: z.string().describe(workspaceIdDescription),
-        command: z
-          .string()
-          .describe(
-            `Shell command to run. Must not create or modify project files; use ${toolNames.edit} or ${toolNames.write} for file changes.`,
-          ),
-        workingDirectory: z
-          .string()
-          .optional()
-          .describe(
-            "Optional working directory relative to the workspace root. Defaults to the workspace root.",
-          ),
-        timeout: z
-          .number()
-          .positive()
-          .max(300)
-          .optional()
-          .describe("Timeout in seconds. Defaults to 30, max 300."),
-      },
-      outputSchema: resultOutputSchema(),
-      ...toolWidgetDescriptorMeta(config, "shell"),
-      annotations: SHELL_TOOL_ANNOTATIONS,
-    },
-    async ({ workspaceId, workingDirectory, ...input }) => {
-      const startedAt = performance.now();
-      const workspace = workspaces.getWorkspace(workspaceId);
-      const cwd = workspaces.resolveWorkingDirectory(
-        workspace,
-        workingDirectory,
-      );
-      const response = await runShellTool(input, {
-        cwd,
-        root: workspace.root,
-      });
-
-      if (response.isError) {
-        logFailedToolResponse(
-          config,
-          {
-            tool: toolNames.shell,
-            workspaceId,
-            workingDirectory: workingDirectory ?? ".",
-            command: input.command,
-            commandLength: input.command.length,
-          },
-          response.content,
-          startedAt,
-        );
-        return response;
-      }
-
-      const summary = {
-        command: input.command,
-        workingDirectory: workingDirectory ?? ".",
-        ...textSummary(response.content),
-      };
-      logToolCall(config, {
-        tool: toolNames.shell,
-        workspaceId,
-        workingDirectory: workingDirectory ?? ".",
-        command: input.command,
-        commandLength: input.command.length,
-        success: true,
-        durationMs: Math.round(performance.now() - startedAt),
-      });
-
-      return {
-        ...response,
-        _meta: {
-          tool: toolNames.shell,
-          card: {
-            workspaceId,
-            path: workingDirectory,
-            summary,
-            payload: { content: response.content },
-          },
-        },
-        structuredContent: {
-          result: contentText(response.content),
-        },
-      };
-    },
-  );
-}
